requirements.txt
scikit-learn
matplotlib
python-dotenv
keyring
pandas
numpy
IPython
requests
google-genai
